Fix all services to match the new models

+++ src/services/commentsService.ts
@@ -1,33 +1,76 @@
 import Comment from '@/models/commentModel';
+import listingModel from '@/models/listingModel';
 import { Comment as CommentType } from '@/types/comment';
 
 const getComments = async () => {
-  return await Comment.find();
+  return await Comment.find().populate('author', 'firstName lastName imageUrl');
 };
 
-const createComment = async (commentData: Partial<CommentType>) => {
-  const newComment = new Comment(commentData);
-  return await newComment.save();
+const createComment = async (
+  listingId: string,
+  authorId: string,
+  commentText: string
+) => {
+  const newComment = await Comment.create({
+    listingId,
+    commentText,
+    author: authorId,
+  });
+
+  await listingModel.findByIdAndUpdate(
+    listingId,
+    { $push: { comments: newComment._id } },
+    { new: true }
+  );
+
+  return await newComment.populate('author', 'firstName lastName imageUrl');
 };
 
 const getCommentById = async (id: string) => {
-  return await Comment.findById(id);
-};
-
-const getCommentsByPostId = async (postId: string) => {
-  return await Comment.find({ postId: postId }).sort({ createdAt: -1 });
-};
-
-const updateComment = async (id: string, updateData: Partial<CommentType>) => {
-  return await Comment.findByIdAndUpdate(
-    id, 
-    { $set: updateData }, 
-    { new: true, runValidators: true } 
+  return await Comment.findById(id).populate(
+    'author',
+    'firstName lastName imageUrl'
   );
 };
 
-const deleteComment = async (id: string) => {
-  return await Comment.findByIdAndDelete(id);
+const getCommentsByPostId = async (listingId: string) => {
+  return await Comment.find({ listingId })
+    .sort({ createdAt: -1 })
+    .populate('author', 'firstName lastName imageUrl');
+};
+
+type UpdateCommentPayload = Pick<CommentType, 'commentText'>;
+
+const updateComment = async (
+  id: string,
+  authorId: string,
+  updateData: UpdateCommentPayload
+) => {
+  const comment = await Comment.findById(id);
+
+  if (!comment) throw new Error('Comment not found');
+  if (comment.author.toString() !== authorId) throw new Error('Unauthorized');
+
+  return await Comment.findByIdAndUpdate(
+    id,
+    { $set: updateData },
+    { new: true, runValidators: true }
+  ).populate('author', 'firstName lastName imageUrl');
+};
+
+const deleteComment = async (id: string, authorId: string) => {
+  const comment = await Comment.findById(id);
+
+  if (!comment) throw new Error('Comment not found');
+  if (comment.author.toString() !== authorId) throw new Error('Unauthorized');
+
+  await Comment.findByIdAndDelete(id);
+
+  await listingModel.findByIdAndUpdate(comment.listingId, {
+    $pull: { comments: id },
+  });
+
+  return { success: true, deletedCommentId: id };
 };
 
 export default {
@@ -36,5 +79,5 @@ export default {
   getCommentById,
   getCommentsByPostId,
   updateComment,
-  deleteComment
+  deleteComment,
 };
